+++ src/components/date-picker/date-picker.vue
@@ -8,7 +8,7 @@
       :triggers="[]"
       :popper-hide-triggers="[]"
       :auto-hide="false"
-      :disabled="props.disabled || props.readonly"
+      :disabled="isDatePickerPopperDisabled"
       :style="{
         width: props.width,
       }"
@@ -22,40 +22,42 @@
             <input
               ref="monthInputRef"
               v-model="monthInput"
-              :class="[datePickerClasses.datePickerInputClasses, 'spr-w-[36px] spr-min-w-[36px]']"
+              :class="getDatePickerInputClasses('36px')"
               type="text"
               placeholder="MMM"
               maxlength="3"
               :disabled="props.disabled"
               :readonly="props.readonly"
-              @input="handleMonthInput(monthInput, $event)"
-              @keyup="handleMonthInput(monthInput, $event)"
+              @input="handleMonthInput(monthInput, $event as KeyboardEvent)"
+              @keyup="handleMonthInput(monthInput, $event as KeyboardEvent)"
+              @keydown="handleMonthInput(monthInput, $event as KeyboardEvent)"
+              @blur="handleMonthInput(monthInput, $event as FocusEvent)"
             />
             <span class="spr-text-color-strong spr-font-size-200 spr-text-color-weak">/</span>
             <input
               ref="dateInputRef"
               v-model="dateInput"
-              :class="[datePickerClasses.datePickerInputClasses, 'spr-w-[20px] spr-min-w-[20px]']"
+              :class="['spr-text-center', getDatePickerInputClasses('20px')]"
               type="text"
               placeholder="DD"
               maxlength="2"
               :disabled="props.disabled"
               :readonly="props.readonly"
-              @input="handleDateInput(dateInput, null, null, $event)"
-              @keyup="handleDateInput(dateInput, null, null, $event)"
+              @input="handleDateInput(dateInput, null, null, $event as KeyboardEvent)"
+              @keyup="handleDateInput(dateInput, null, null, $event as KeyboardEvent)"
             />
             <span class="spr-text-color-strong spr-font-size-200 spr-text-color-weak">/</span>
             <input
               ref="yearInputRef"
               v-model="yearInput"
-              :class="[datePickerClasses.datePickerInputClasses, 'spr-w-[42px] spr-min-w-[42px]']"
+              :class="getDatePickerInputClasses('42px')"
               type="text"
               placeholder="YYYY"
               maxlength="4"
               :disabled="props.disabled"
               :readonly="props.readonly"
-              @input="handleYearInput(yearInput, $event)"
-              @keyup="handleYearInput(yearInput, $event)"
+              @input="handleYearInput(yearInput, $event as KeyboardEvent)"
+              @keyup="handleYearInput(yearInput, $event as KeyboardEvent)"
             />
           </div>
           <div class="spr-flex spr-items-center spr-justify-center">
@@ -75,12 +77,7 @@
             <!-- Tabs -->
             <div class="spr-flex spr-gap-1">
               <spr-button
-                :class="[
-                  'spr-cursor-pointer',
-                  {
-                    'spr-background-color-pressed !spr-shadow-button': currentTab === 'tab-months',
-                  },
-                ]"
+                :class="getTabClasses('tab-months')"
                 variant="secondary"
                 size="small"
                 @click="currentTab = 'tab-months'"
@@ -88,12 +85,7 @@
                 {{ getMonthObject('monthValue', calendarTabPageData.selectedMonth)?.fullText }}
               </spr-button>
               <spr-button
-                :class="[
-                  'spr-cursor-pointer',
-                  {
-                    'spr-background-color-pressed !spr-shadow-button': currentTab === 'tab-years',
-                  },
-                ]"
+                :class="getTabClasses('tab-years')"
                 variant="secondary"
                 size="small"
                 @click="currentTab = 'tab-years'"
@@ -251,12 +243,22 @@
         </div>
       </template>
     </Menu>
-    <p v-if="props.displayHelper" :class="datePickerClasses.datePickerInputHelperClasses">
-      <slot name="helperMessage">
-        <Icon v-if="props.helperIcon" :icon="props.helperIcon" width="20px" height="20px" />
-        <span>{{ props.helperText }}</span>
-      </slot>
-    </p>
+    <template v-if="datePickerErrors.length > 0">
+      <p :class="datePickerClasses.datePickerInputHelperClasses">
+        <slot name="helperMessage">
+          <Icon icon="ph:warning-circle-fill" width="20px" height="20px" />
+          <span>{{ datePickerErrors[0].message }}</span>
+        </slot>
+      </p>
+    </template>
+    <template v-else>
+      <p v-if="props.displayHelper" :class="datePickerClasses.datePickerInputHelperClasses">
+        <slot name="helperMessage">
+          <Icon v-if="props.helperIcon" :icon="props.helperIcon" width="20px" height="20px" />
+          <span>{{ props.helperText }}</span>
+        </slot>
+      </p>
+    </template>
   </div>
 </template>
 
@@ -288,6 +290,7 @@ const {
   monthsList,
   monthInput,
   yearInput,
+  datePickerErrors,
   calendarTabPageData,
   calendarTabIsMinMonth,
   calendarTabIsMaxMonth,
@@ -300,6 +303,9 @@ const {
   calendarTabIsDateIsDisabled,
   calendarTabHandleDateInput,
   getMonthObject,
+  getDatePickerInputClasses,
+  getTabClasses,
+  isDatePickerPopperDisabled,
   calendarTabPrevMonth,
   calendarTabNextMonth,
   monthTabHandleSelectedMonth,