+++ src/components/input/input.vue
@@ -22,25 +22,25 @@
       <div v-if="$slots.icon" :class="inputClasses.iconSlotClasses">
         <slot name="icon" />
       </div>
-      <label
-        v-if="props.displayHelper"
-        :class="[
-          inputClasses.helperClasses,
-          'spr-font-size-200 spr-font-line-height-400 spr-font-normal',
-          'spr-flex spr-items-center spr-gap-size-spacing-5xs',
-          'spr-mt-size-spacing-4xs',
-        ]"
-      >
-        <slot name="helperMessage">
-          <icon v-if="props.helperIcon" :icon="props.helperIcon" width="20px" height="20px" />
-          {{ props.helperText }}
-        </slot>
-      </label>
     </div>
+
+    <label
+      v-if="props.displayHelper"
+      :class="[
+        inputClasses.helperClasses,
+        'spr-font-size-200 spr-font-line-height-400 spr-font-normal',
+        'spr-flex spr-items-center spr-gap-size-spacing-5xs',
+      ]"
+    >
+      <slot name="helperMessage">
+        <icon v-if="props.helperIcon" :icon="props.helperIcon" width="20px" height="20px" />
+        {{ props.helperText }}
+      </slot>
+    </label>
   </div>
 </template>
 
-<script lang="ts" setup>
+<script setup lang="ts">
 import { useSlots } from 'vue';
 import { Icon } from '@iconify/vue';
 import { inputPropTypes, inputEmitTypes } from './input';
@@ -54,7 +54,7 @@ const slots = useSlots();
 const { inputClasses, onInput } = useInput(props, slots, emit);
 </script>
 
-<style lang="scss" scoped>
+<style scoped>
 .number-input::-webkit-outer-spin-button,
 .number-input::-webkit-inner-spin-button {
   margin: 0;